refactor: split parser.go into domain-specific files
Move parse methods out of the 1200-line parser.go into files
grouped by domain object:

- parse_soldier.go: ParseSoldier, ParseSoldierState
- parse_vehicle.go: ParseVehicle, ParseVehicleState
- parse_events.go: ParseProjectileEvent, ParseGeneralEvent,
  ParseKillEvent, ParseFpsEvent, ParseTimeState
- parse_ace3.go: ParseAce3DeathEvent, ParseAce3UnconsciousEvent
- parse_chat.go: ParseChatEvent, ParseRadioEvent
- parse_marker.go: ParseMarkerCreate, ParseMarkerMove, ParseMarkerDelete

parser.go retains the Parser struct, constructor, helpers,
MissionContext, and ParseMission. Pure file reorganization,
no behavior change.
